+++ shong/chapter4/src/main/java/com/example/chapter4/domain/store/entity/Area.java
@@ -6,9 +6,8 @@ import lombok.*;
 import java.util.ArrayList;
 import java.util.List;
 
-/**
- * 지역 정보를 나타내는 엔티티입니다.
- */
+//지역 정보를 나타내는 엔티티입니다.
+
 @Entity
 @Getter
 @Builder
